Recording options, change to command structure, and silent create
Adding several different recording options including the option to delete a command and quit without saving. Changed options already in place for record and tidied it up. Also changed the command structure. Calling batch files without the .bat extension sometimes causes them to not execute properly. To fix this, commands now have two batch files associated with them. One for calling the command (located as before in the Alix home directory) and one for actually executing the command (located in the new cmds directory). The batch file located in the upper dirctory calls the cmds directory batch file with the .bat extension allowing for better commands. Final change allowed alix_create to be silent using option so that deleting a command was cleaner.