Adding automatic command updating and fixing argument passing bug
Added an extra lastmod argument to the stored alix dictionary which allows Alix to detect at what time a command file has been previously modified. When the stored lastmod time is not equivalent to the modtime as specified by the os Alix will now read the file and update the dictionary stored command. This allows alix to properly display the correct cmd when asked to list or show commands with the verbose setting turned on.

This commit also fixes a bug which would cause newly created commands to not the needed batch argument passing parameter in the upper level script, so that arguments could be passed through alix commands.